+++ src/util/enumcollect/WaybillTypeEnum.java
@@ -1,4 +1,4 @@
-package util;
+package util.enumcollect;
 
 public enum WaybillTypeEnum {
     // 일반
